- Add copy button and copy code block context menu to BotChatBubble
- Refactor chat view into smaller components for better code maintenance
