+++ OLMoE.swift/Constants/Localizable.xcstrings
@@ -26,6 +26,12 @@
     "Cancel" : {
 
     },
+    "Copied" : {
+
+    },
+    "Copy Code" : {
+
+    },
     "Delete" : {
 
     },
